cmd: fix send Redirect packet and encode/decode size as a varint in Redirect packet
This way an encoded Redirect Packet of more than 127 bytes can be sent
as a delimited protobuf message. To encode the size of a delimited
encoded Redirect packet of more than 127 bytes more than one byte is
needed.

There were several errors in the function that implements the sending of
the Redirect packet.